[25.1] Fix max http proxy memory limit commit_hash:893d02e6bb67419ab99e78b07ab71681cc29c4e5

+++ yt/yt/server/http_proxy/config.h
@@ -190,8 +190,8 @@ class TMemoryLimitsConfig
     : public NYTree::TYsonStruct
 {
 public:
-    std::optional<i32> Total;
-    std::optional<i32> HeavyRequest;
+    std::optional<i64> Total;
+    std::optional<i64> HeavyRequest;
 
     REGISTER_YSON_STRUCT(TMemoryLimitsConfig);
